requirements.txt
requests==2.28.1
demucs==4.0.0
requests==2.28.1
hikari==2.0.0.dev115
pydub==0.25.1
hikari-lightbulb==2.3.1
yt_dlp==2023.3.4
hikari-miru==2.0.4
pyyaml==6.0
python_dotenv==1.0.0